Feature #314 - add window confirm for file deletion (#460)
* add window confirm for file deletion

* remove lint error for confirm

+++ client/src/app/components/App/Modal/PagesList/PageRow/PageRow.jsx
@@ -27,7 +27,9 @@ function collect(_connect, monitor) {
 class PageRow extends Component {
   deletePage = (e) => {
     e.stopPropagation();
-    this.props.deletePage(this.props.page._id);
+    if (confirm('Are you sure you want to delete this file?')) { // eslint-disable-line no-restricted-globals
+      this.props.deletePage(this.props.page._id);
+    }
   }
 
   duplicatePage = (e) => {
